requirements.txt
appdirs==1.4.4
asgiref==3.3.4
boto3==1.17.105
botocore==1.20.105
cfgv==3.3.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
gunicorn==20.1.0
identify==2.2.10
jmespath==0.10.0
packaging==21.0
Pillow==8.2.0
pip-autoremove==0.9.1
pip-check-reqs==2.2.2
pre-commit==2.13.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
virtualenv==20.4.7
whitenoise==5.2.0
psycopg2-binary==2.8.6
